Add changes to webpack settings
