+++ src/index.html
@@ -10,7 +10,7 @@
 <body>
   <div class="root">
     <header class="header root__section">
-      <img src="./images/logo.svg" alt="mesto logo" class="logo">
+      <img src="<%=require('../images/logo.svg')%>" alt="mesto logo" class="logo">
     </header>
     <div class="profile root__section">
       <div class="user-info">
@@ -42,7 +42,7 @@
     </div>
     <div id="popup-add" class="popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close popup__add">
+        <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close popup__add">
         <h3 class="popup__title">Новое место</h3>
         <form class="popup__form" name="new">
             <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Название" minlength="2" maxlength="30" required>
@@ -56,7 +56,7 @@
 
     <div id="popup-edit" class="popup">
       <div class="popup__content">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close">
         <h3 class="popup__title">Редактировать профиль</h3>
         <form class="popup__form" name="self">
             <input type="text" name="name" class="popup__input popup__input_type_name" placeholder="Имя" minlength="2" maxlength="30" required>
@@ -70,7 +70,7 @@
 
     <div id="popup-image" class="popup">
       <div class="popup__picture">
-        <img src="./images/close.svg" alt="" class="popup__close">
+        <img src="<%=require('../images/close.svg')%>" alt="" class="popup__close">
         <img class="popup__image">
       </div>
     </div>
