+++ webpack.config.js
@@ -24,10 +24,9 @@ module.exports = {
             {
                 test: /\.(png|jpg|gif|ico|svg)$/,
                 use: [
-                     'file-loader?name=../images/[name].[ext]', // указали папку, куда складывать изображения
+                     'file-loader?name=./images/[name].[ext]', // указали папку, куда складывать изображения
                      {
                          loader: 'image-webpack-loader'
-                         
                      },
                     ],
             },
